🛡️ Guardian (JULES): Consolidated DiffView components into UI library
WHAT:
- Merged src/components/book/elements/diff-view.tsx into src/components/ui/diff-view.tsx
- Updated MDX files to use SideBySideDiff
- Deleted obsolete duplicate file
- Updated barrel files

WHY:
- Duplicate logic across book and UI components
- Inconsistent maintenance and duplicated diffing algorithms

JULES CONTEXT:
- Verified no Autonomous task conflicts in .jules/task-log.md
- Session logged: .Jules/guardian/2026-08-06/

IMPACT:
- Files: reduced by 1
- Code duplication reduced

VERIFICATION:
- All tests pass
- TypeScript check: ✓
- Linting: ✓
- Build: ✓
- JULES compliance: ✓
