+++ .jules/task-log.md
@@ -22,3 +22,9 @@
 - Merged `src/components/ide/utils.ts` into `src/lib/format.ts`
 - Updated import in `src/components/ide/prompt-ide.tsx`
 - Deleted `src/components/ide/utils.ts`
+
+## 2026-08-06 16:50 - [GUARDIAN] Session started
+
+- Directory: .Jules/guardian/2026-08-06
+- Phase: PRE-FLIGHT
+- JULES Check: COMPLETE
