+++ src/content/book/08-iterative-refinement.mdx
@@ -19,42 +19,42 @@ Most prompt failures fall into a handful of categories. Learning to recognize th
 
 One of the most common issues. Without explicit constraints, models tend to be thorough rather than concise.
 
-<DiffView
-  before="Explain how photosynthesis works."
-  after={`Explain how photosynthesis works in 3-4 sentences suitable for a 10-year-old.`}
-  beforeLabel="Original"
-  afterLabel="Refined"
+<SideBySideDiff
+  original="Explain how photosynthesis works."
+  modified={`Explain how photosynthesis works in 3-4 sentences suitable for a 10-year-old.`}
+  originalLabel="Original"
+  modifiedLabel="Refined"
 />
 
 ### Problem: Output Too Vague
 
 Vague prompts produce vague outputs. The model can't read your mind about what "better" means or which aspects matter most to you.
 
-<DiffView
-  before="Give me tips for better presentations."
-  after={`Give me 5 specific, actionable tips for improving technical presentations to non-technical stakeholders. For each tip, include a concrete example.`}
-  beforeLabel="Original"
-  afterLabel="Refined"
+<SideBySideDiff
+  original="Give me tips for better presentations."
+  modified={`Give me 5 specific, actionable tips for improving technical presentations to non-technical stakeholders. For each tip, include a concrete example.`}
+  originalLabel="Original"
+  modifiedLabel="Refined"
 />
 
 ### Problem: Wrong Tone
 
 Tone is subjective and varies by context. What sounds "professional" to the model might not match your organization's voice or the relationship with your recipient.
 
-<DiffView
-  before="Write an apology email for missing a deadline."
-  after={`Write a professional but warm apology email for missing a project deadline. The tone should be accountable without being overly apologetic. Include a concrete plan to prevent future delays.`}
-  beforeLabel="Original"
-  afterLabel="Refined"
+<SideBySideDiff
+  original="Write an apology email for missing a deadline."
+  modified={`Write a professional but warm apology email for missing a project deadline. The tone should be accountable without being overly apologetic. Include a concrete plan to prevent future delays.`}
+  originalLabel="Original"
+  modifiedLabel="Refined"
 />
 
 ### Problem: Missing Key Information
 
 Open-ended requests get open-ended responses. If you need specific types of feedback, you must ask for them explicitly.
 
-<DiffView 
-  before="Review this code."
-  after={`Review this Python code for:
+<SideBySideDiff
+  original="Review this code."
+  modified={`Review this Python code for:
 1. Bugs and logical errors
 2. Performance issues
 3. Security vulnerabilities
@@ -63,17 +63,17 @@ Open-ended requests get open-ended responses. If you need specific types of feed
 For each issue found, explain the problem and suggest a fix.
 
 [code]`}
-beforeLabel="Original"
-afterLabel="Refined"
+originalLabel="Original"
+modifiedLabel="Refined"
 />
 
 ### Problem: Inconsistent Format
 
 Without a template, the model will structure each response differently, making comparison difficult and automation impossible.
 
-<DiffView 
-  before="Analyze these three products."
-  after={`Analyze these three products using this exact format for each:
+<SideBySideDiff
+  original="Analyze these three products."
+  modified={`Analyze these three products using this exact format for each:
 
 ## [Product Name]
 
@@ -84,8 +84,8 @@ Without a template, the model will structure each response differently, making c
 **Rating:** X/10
 
 [products]`}
-beforeLabel="Original"
-afterLabel="Refined"
+originalLabel="Original"
+modifiedLabel="Refined"
 />
 
 ## Systematic Refinement Approach
